+++ doc/versionHistory.rst
@@ -6,6 +6,16 @@
 Version History
 ##################
 
+.. _lsst.ts.wep-15.0.0:
+
+-------------
+15.0.0
+-------------
+
+* Fix CalcZernikesUnpairedTask to handle new `numCores` parameter in estimateZernikes.
+* Redefine Unpaired tasks to run on each detector independently.
+* Remove cutOutDonutsUnpairedTask since it replicates code in cutOutDonutsCwfsTask.
+
 .. _lsst.ts.wep-14.17.2:
 
 -------------